refactor: 헤더 위치 이동 및 스크롤 시 애니메이션을 커스텀 훅으로 분리
- ref: #33
