+++ src/pages/Home/index.tsx
@@ -1,5 +1,22 @@
+import { useSetAtom } from 'jotai';
+import Header from '@components/Header';
+import useScrollVisibility from '@hooks/useScrollVisibility';
+import { headerVisibilityAtom } from '@stores/layoutState';
+
 const Home = () => {
-  return <div />;
+  const setHeaderVisibility = useSetAtom(headerVisibilityAtom);
+
+  useScrollVisibility(setHeaderVisibility, 10);
+
+  return (
+    <div>
+      <Header
+        onSearchClick={() => {}}
+        onNotificationClick={() => {}}
+        onCartClick={() => {}}
+      />
+    </div>
+  );
 };
 
 export default Home;
