+++ vite.config.ts
@@ -14,6 +14,7 @@ export default defineConfig({
         find: '@components',
         replacement: path.resolve(__dirname, 'src/components'),
       },
+      { find: '@hooks', replacement: path.resolve(__dirname, 'src/hooks') },
       { find: '@mocks', replacement: path.resolve(__dirname, 'src/mocks') },
       { find: '@pages', replacement: path.resolve(__dirname, 'src/pages') },
       { find: '@stores', replacement: path.resolve(__dirname, 'src/stores') },
